+++ plugins/gs/src/index.ts
@@ -6,4 +6,9 @@ export {
 } from './plugin';
 export { gsAuthApiRef, GSAuth } from './apis';
 export { ProviderSettings as GSProviderSettings } from './components/ProviderSettings';
-export { isEntityDeploymentsAvailable as isEntityGSDeploymentsAvailable } from './components/utils/entity';
+export {
+  isEntityDeploymentsAvailable as isEntityGSDeploymentsAvailable,
+  isEntityLatestReleaseAvailable as isEntityGSLatestReleaseAvailable,
+  isEntityHelmChartsAvailable as isEntityGSHelmChartsAvailable,
+} from './components/utils/entity';
+export { columnFactories as GSColumnFactories } from './components/catalog/columns';